requirements.txt
anyio==3.7.1
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.7.14
chardet==3.0.4
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
Django==5.2
django-appconf==1.1.0
django-ckeditor-5==0.2.18
django-cors-headers==4.7.0
django-dbbackup==4.2.1
django-extensions==3.2.3
django-filter==25.1
django-imagekit==5.0.0
django-js-asset==3.1.2
django-redis==6.0.0
django-seed==0.3.1
django-storages==1.14.4
django-unfold==0.70.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.28.0
Faker==37.3.0
fuzzywuzzy==0.18.0
googletrans==4.0.0rc1
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==3.10
imagekitio==4.1.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Levenshtein==0.26.1
libretranslatepy==2.1.1
lxml==5.3.0
packaging==24.2
pilkit==3.0
pillow==11.2.1
PyJWT==2.6.0
python-dotenv==1.1.1
python-Levenshtein==0.26.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.11.0
redis==6.2.0
referencing==0.36.2
requests==2.32.4
requests-toolbelt==0.10.1
rfc3986==1.5.0
rpds-py==0.25.1
setuptools==79.0.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
toposort==1.10
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
waitress==3.0.2
whitenoise==6.8.2
